feat : add base and surface

+++ TP2/TP2/triangle.h
@@ -23,7 +23,7 @@ public:
 
     float Base();
     float Hauteur();
-    double Surface();
+    float Surface();
     float Longueurs();
 
     bool isIsosceles();
